[CREATURE]: writing general functions to get/change attribute value

+++ Sources/player.h
@@ -85,7 +85,7 @@ private:
 
 
     // ######## SKILLS
-    std::array<int, Skills::TOTAL> mb_skillLevel;
+    std::array<int, static_cast<int>(Skills::TOTAL)> mb_skillLevel;
